+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Deprecate `service/defaultcomponents` go package (#4622)
 
+## 💡 Enhancements 💡
+
+- Transform configmapprovider.Retrieved interface to a struct (#4789)
+
 ## v0.44.0 Beta
 
 ## 🛑 Breaking changes 🛑